refactor: rename XDuinoRails to XDuinoRails_MotorControl
Renames the main library from `XDuinoRails` to `XDuinoRails_MotorControl` for improved clarity and consistency.

This change includes:
- Renaming the main library directory.
- Updating the `library.json` manifest.
- Renaming the main header file to `XDuinoRails_MotorControl.h`.
- Updating all example sketches and `platformio.ini` build environments to use the new library name and header.
