+++ examples/FindMotorConfig/FindMotorConfig.cpp
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include <XDuinoRails.h>
+#include <XDuinoRails_MotorControl.h>
 
 // Pin Definitions
 const int pwmAPin = D7;
