+++ lib/XDuinoRails_MotorControl/library.json
@@ -1,5 +1,5 @@
 {
-  "name": "XDuinoRails",
+  "name": "XDuinoRails_MotorControl",
   "version": "0.0.1",
   "description": "A library for controlling DC motors with BEMF feedback.",
   "keywords": "motor, bemf, rp2040",
@@ -11,7 +11,7 @@
   ],
   "license": "AGPL-3.0-or-later",
   "headers": [
-    "XDuinoRails.h",
+    "XDuinoRails_MotorControl.h",
     "XDuinoRails_MotorDriver.h",
     "pi_controller.h"
   ],
