+++ platformio.ini
@@ -40,7 +40,7 @@ framework = arduino
 src_dir = examples/AdvancedMovement
 lib_deps =
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
     adafruit/Adafruit NeoPixel
     mathertel/RotaryEncoder
 
@@ -66,7 +66,7 @@ framework = arduino
 src_dir = examples/STM32_GettingStarted
 lib_deps =
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
     adafruit/Adafruit NeoPixel
 
 [env:find_motor_config]
@@ -77,7 +77,7 @@ src_dir = examples/FindMotorConfig
 lib_deps =
     adafruit/Adafruit NeoPixel
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
 
 [env:getting_started]
 platform = https://github.com/maxgerhardt/platform-raspberrypi.git
@@ -86,7 +86,7 @@ framework = arduino
 src_dir = examples/GettingStarted
 lib_deps =
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
     adafruit/Adafruit NeoPixel
 
 [env:advanced]
@@ -96,7 +96,7 @@ framework = arduino
 src_dir = examples/Advanced
 lib_deps =
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
     adafruit/Adafruit NeoPixel
 
 [env:rotary_encoder_control]
@@ -106,6 +106,6 @@ framework = arduino
 src_dir = examples/RotaryEncoderControl
 lib_deps =
     denyssene/SimpleKalmanFilter
-    XDuinoRails
+    XDuinoRails_MotorControl
     adafruit/Adafruit NeoPixel
     mathertel/RotaryEncoder
